refactor: migrate column config slider to typescript
